requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-jose
passlib
python-dotenv
pydantic-settings
pydantic[email]
openai
boto3
botocore
python-multipart
passlib